requirements.txt
pymongo==3.5.1
redis==3.5.3
grpcio
grpcio-tools